requirements.txt
flask
gunicorn
python-dotenv
vaderSentiment
groq
